re-arrange the files in utils
