+++ pyanomaly/utils/__init__.py
@@ -1,3 +1,3 @@
-from .cmd import parse_args
-from .system import system_setup
-from .utils import create_logger, get_tensorboard
+from .system import system_setup, parse_args
+from .recorders import create_logger, get_tensorboard
+from .tools import *
